+++ tests/fermi/test_fermionic.py
@@ -17,9 +17,8 @@ from copy import copy, deepcopy
 
 import pytest
 
-from pennylane.fermi.fermionic import FermiWord, FermiC, FermiA
+from pennylane.fermi.fermionic import FermiWord
 
-# pylint: disable=too-few-public-methods
 
 fw1 = FermiWord({(0, 0): "+", (1, 1): "-"})
 fw2 = FermiWord({(0, 0): "+", (1, 0): "-"})
@@ -173,49 +172,3 @@ class TestFermiWord:
         """Test that an error is raised if the operator orders are not correct."""
         with pytest.raises(ValueError, match="The operator indices must belong to the set"):
             FermiWord(operator)
-
-
-class TestFermiC:
-    """Test the methods of the creation operator FermiC"""
-
-    @pytest.mark.parametrize("wire", [1, 3])
-    def test_initialization(self, wire):
-        """Test __init__ function returns the expected FermiWord object"""
-        op = FermiC(wire)
-        op_dict = {(0, wire): "+"}
-
-        assert isinstance(op, FermiWord)
-        assert len(op) == 1
-        assert list(op.keys()) == [(0, wire)]
-        assert list(op.values()) == ["+"]
-
-        assert op.sorted_dic == op_dict
-
-    @pytest.mark.parametrize("wire", ["a", -2, [0, 1], 1.2])
-    def test_bad_wire_raises_error(self, wire):
-        """Test that passing a value for wire that is not a positive integer raises an error"""
-        with pytest.raises(ValueError, match="expected a single, positive integer value for wire"):
-            _ = FermiC(wire)
-
-
-class TestFermiA:
-    """Test the methods of the annihilation operator FermiA"""
-
-    @pytest.mark.parametrize("wire", [1, 3])
-    def test_initialization(self, wire):
-        """Test __init__ function returns the expected FermiWord object"""
-        op = FermiA(wire)
-        op_dict = {(0, wire): "-"}
-
-        assert isinstance(op, FermiWord)
-        assert len(op) == 1
-        assert list(op.keys()) == [(0, wire)]
-        assert list(op.values()) == ["-"]
-
-        assert op.sorted_dic == op_dict
-
-    @pytest.mark.parametrize("wire", ["a", -2, [0, 1], 1.2])
-    def test_bad_wire_raises_error(self, wire):
-        """Test that passing a value for wire that is not a positive integer raises an error"""
-        with pytest.raises(ValueError, match="expected a single, positive integer value for wire"):
-            _ = FermiA(wire)
